Added some changes to the guest approach
Instead of a separate table for guest replies, they are now stored in the same table as every other reply. The user id is set to zero, and the user-provided info such as name and email are stored in a separate table. I also added a little security log when a guest user posts.

Furthermore, fixed a little bug where it wasn't possible to remove a reported forum post.: